README.md: Use consistent style for code snippets
The only reason to use the long form for code snippets is if you
want to tell the syntax highlighter how to interpret your code. It
doesn't work very well for `console` snippets, so it's better to
just use the short form.

Also, prefix the shell lines with `$` to be more obvious that they
are for the shell.

Signed-off-by: Alan Orth <alan.orth@gmail.com>